requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
geopy==2.4.1
openai==1.3.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.2
celery==5.3.4
redis==5.0.1
# psycopg2-binary>=2.9.9  # Optional: Only needed for PostgreSQL. Install separately if using PostgreSQL
Pillow>=10.1.0
django-extensions==3.2.3
setuptools>=68.0.0
